Refactor image creation so it uses packer.
Keeping around the old create_gce_image script for the time being for
existing users, but would like to delete it.

Migrated tarball creation into its own script. This is needed for C2D.
The packer scripts are extendable to support AWS (different PR).
Packer does most of the work, the scripts just facilitate configuring
the packer template variables.

Added gcloud installation to the developer bootstrap because the
gcloud shipped in GCE is too old for our usage patterns. I dont
force the install because gcloud installation has a lot of prompting,
which is annoying if people have no interest in using GCE.