Updated remainder of development scripts.
Some of these are about to become obsolete, but this clears my mental backlog
and stuff I did last week that hadnt gone through review.

General theme is to remove google bias and provide some aws support.
There are a few CLs here together, but original testing was time consuming
across all the platforms. This is a backport of heuristics that were tested,
but the backport itself is only tested on GCE (both old image creation and
new packer image creation) from create_dev to a new vm all the way through
building and deploying a release.

ok