Adds in migration to new configuration.
Add support to install/start/run with new config.

Brings up spinnaker in prod, assuming the /opt files have added
the spring.config.location

Add config/ to spring path in production and gradle

For this to work in dev, you need to say
   ../spinnaker/google/dev/refresh_sources.sh --nobuild --hack_gradle
from the build directory to change the spring path in the gradle files.
This is temporary until the files themselves are modified to accomodate
the new strategy.

The --hack_gradle is not required for production enablement.
Either old or new style is available for any release.
Except this CL requires deck be manually configured to use the new style.