Convert deck settings.js to new config.
This requires reconfigure to inject the yaml settings into the javascript file.
To make this more manageable and clearer, there is a variable block at the
top of the file that is used to inject the bindings. The downside is that
this requires the variables be explicitly called out. The upside is that
the file maintains its integrity and can be regenerated.

Removed the hack from the gradle scripts, but clouddriver does not yet
start using the new configuration on gradle (but everything does work
on the production install).