chore(buildtool): implementation cleanup
Changed strategy for dealing with nebula friction by using a decoy origin
repository rather than removing all the confusing tags and versions.
Using the nebula release.verison switch eliminates the need to remove tags,
however it insists on pushing a new tag (and the wrong pattern at that).
The decoy repository tricks nebula into pushing the tag to somewhere irrelvant.

Cleaned up some of the gradle build command classes, adding build_rpms.
Added handler to try to detect causes of failures. So far, it just recognizes
bintray issues. More will be added as they come up.

There's a lot of diff resulting from moving the BuildHalyardCommand further
down in the file. That was because the command is now a refinement of
building debians since they share a lot of commonality.

Added additional logging around threading to help debug a deadlock if
there is one there.