chore(buildtool): Be efficient at doing nothing
There are three optimizations here.

First is a change to build_bom that, when refreshing an existing bom,
will only refresh the bom if the repository commits have changed.
This makes it transparent to maintain a bom in conjunction with the git
repos (the primitive command requires you explicitly set a parameter to
do this, however the flow script sets the parameter by default unless you
tell it not to, such as for a new branch).

Second, given that boms are easily maintained with source changes, the
"flow_build" script starts a build using the bom from the previous build.
This also adds an interim local bom for "unbuilt" artifacts so that failed
builds can continue where they left off. The next build will still pick up
on any repository changes (e.g. a fix to the build failure).

Third, given that this makes it common for most repositories to offer no
work, the RepositoryCommands check the remote repo for changes before even
cloning it. While the clone isnt that expensive, this also cuts down on the
semver analysis and all the debugging logging noise.