fix(citest): Fix operation polling in kube v2 tests (#3172)
A few of the kube v2 tests use a status class of
SynchronousHttpOperationStatus when sending a request to trigger
a pipeline; this causes the test to return success immediately
and move on to testing the contract condition. We should remove this
and fall back on the default (GateTaskStatus) which will poll for
completion of the pipeline before trying to verify the contract.