fix(deploy): Fix base image used in deploy tests (#27) (#29)
The tests to deploy a server group to GCE take a base image, then
use a startup script to install apache using apt. At some point
recently, the command to install apache started pausing for user
input (despite passing the -y flag to apt-get install), meaning
that the startup script would hang forever, and the health check
waiting for the instance to respond on port 80 would also hang
forever. This was making tests of deploying to GCE fail.

In general, a pre-baked image should have all of the dependencies
already installed. I've made a new image called apache-base-image,
which is just Debian 10 with apache2 already installed and configured.
Let's use this image in the tests instead of installing apache as a
startup script, which is both less hermetic and less similar to the
common use case we'd like to support.

This PR updates the image name and removes the startup script.

Co-authored-by: Eric Zimanyi <ezimanyi@google.com>