fix(kubernetes): Fix race condition in V2 deployment (#33)
When deploying using the V2 provider, control is ceded back to the
CI tests once the deployments have been applied, whether or not they
are stable yet.  This is causing errors on setting up port forwarding,
as the pods are sometimes not running when we try to forward to them.
To address this, let's block on deployment availability before trying
to port forward.