fix(kubernetes): Actually fix race condition in V2 deployment
The last commit added some great logic to make sure that V2
deployments are ready before we try to port-forward to them,
except that it added the logic to the V1 deployer instead of
the V2 deployer. Move this logic to the V2 deployer, and add
a couple of log lines so it's easier to see that it's working.