fix(tests): Add explicit account flag
Some of the calls to gcloud during the integration tests don't
explicitly pass in an account flag. This creates a dependence on
the globally-set account flag that works most of the time but
fails if another Jenkins job happens to change the account at
the wrong time.

Make this more robust by explicitly passing in the --account flag
as we do in most other places.