fix(buildtool): Push changelog command should pull origin (#49) (#51)
I was trying to figure out why any manual changes I made to the raw
changelog (notably deleting the no-longer-needed 1.14, 1.15, 1.16
entries) were disappearing.  It turns out that while we fetch the
origin commit, we don't do anything with it, and commit on top of
what's currently on the jenkins machine, then force push that to
the origin.

Instead, just reset master to origin/master before making any
changes. There should not be any local changes in this directory;
if there are, it would be from an error in some prior run, and we'd
rather just start fresh (as we do in the first block where we just
clone).

Also remove the force push; if the origin *has* changed since we
fetched a few lines above, let's fail rather than overwrite what's
there.

Co-authored-by: Eric Zimanyi <ezimanyi@google.com>