chore(release): Auto-deprecate last patch release
One of the manual steps in doing a patch release of Spinnaker is
to manually run hal admin deprecate ... to deprecate the prior
patch release. Instead of having this as a manual step in the
runbook, just run this automatically after publishing the new
release.

If this is not a patch release (ie, 1.X.0) then this will not
deprecate anything. It will still be a manual step to deprecate
the last patch release of the N-3 minor release when releasing
a new minor release (both by running the hal admin deprecate
command and by marking the changelog as deprecated on spinnaker.io).
But as that only happens once every two months, while we often do
several patch releases a week (over the three supported versions)
this should significantly cut back on toil and human error. (Not
that I have ever accidentally deprecated the wrong release...😬.)