chore(buildtool): Auto-deprecate old Spinnaker versions (#43) (#55)
The publish_changelog function in the buildtool currently
publishes a new changelog for the newly-released Spinnaker version.
After publishing a patch version, the release manager needs to
manually edit the spinnaker.github.io repository to deprecate the
prior patch version, which is unnecessary manual toil.

This change updates the publish_changelog so that, in addition to
publishing a new changelog, it also deprecates the changelog
corresponding to the prior patch release. If this is the first
patch release of a minor version, it will short-circuit and will
not deprecate anything. (Maybe at some point we could update it
to deprecate the last patch release of the N-3 minor version but
for now that will stay manual.)

This also adds a newline at the end of each generated changelog,
as this always gets added when eventually deprecating it (whether
manually or via this script), increasing the diff.

Co-authored-by: Eric Zimanyi <ezimanyi@google.com>