chore(build): Update docs site changelog generation (#48) (#59)
Instead of making a new gist for each patch release, we'll
now use a single top-level gist for each minor release, and
put the release notes for each patch version in a separate
file in the gist (ex: 1.18.0.md, 1.19.1.md).

We'll no longer copy-paste the prior release's full changelog
into each patch release's changes. Instead, we'll just refer
to all prior patch releases by reference in the docs site,
avoiding a lot of copy-paste and duplication.  (And if ever
we needed to amend the release notes for an old version we'd
currently need to do that everywhere it was copy-pasted but
now only in one place.)

This commit updates the code that generates the new file in
spinnaker.github.io to pull in the new changelog. It pulls
in reverse order the file corresponding to every patch release
in the current minor version. It also fixes the fact that we
were incorrectly using a self-closing tag for the script
element which is not allowed.

Co-authored-by: Eric Zimanyi <ezimanyi@google.com>