fix(changelog): Changelog generation should use origin/master
The changelog generation is using the master branch as part of its
logic to compute what changes are part of a particular build; while
we fetch origin/master every time the build runs, we never update
the local master with the changes, so it is currently very out
of date on the Jenkins machine. Rather than rely on this local
branch, just refer to origin/master directly.