chore(mergify): Add mergify config
There was no mergify config for this repo; add one, copying the same template
that is in other repos (including the pending changes to require a release
manager to sign off on release branch changes).

As there does not seem to be any CI set up for the repo, I did not include
that condition in the mergify config, though of course we can add that
as a condition (and likely also at the Github branch level) if we do
ever add that.