chore(containers): don't save the cache until the end
The problem with saving the cache in parallel with building the
containers is that docker tries to include the cache file in the build
context, but the file is still being written as the context is being
prepared. This leads to this error:

  Can't add file /workspace/clouddriver-master.tgz to tar: archive/tar: write too long

It's essentially harmless since we don't want that file in the context
anyway, but still.