chore(buildtool): Add function to get next patch parameters (#80)
The most tedious and error-prone part of doing a Spinnaker patch
release is looking up the prior patch release in the series and
copying over all of the parameters to a new Jenkins job.

In the worst case, forgetting to increment the version leads to
re-releasing on top of an existing Spinnaker version (or so I am
told...😭).

The only user-configurable parameter should be the branch on
which we are doing a patch release; the next patch version can
be inferred from the current patch and all the other parameters
(release name, changelog gist, min halyard version) generally stay
the same.

Add a command to the buidltool that, given a major/minor release,
returns all the data needed to do a patch release. The idea is that
we'll only ask the user to input the major/minor release in Jenkins;
it will then call this function to figure out the rest of the parameters
and then call the Publish_SpinnakerRelease job.

We'll still use Publish_SpinnakerRelease under the hood, and that
job will remain available, both for doing initial x.y.0 releases,
as well as if there's ever a need to change something from what it
was in the last patch release.