feat(gcb): add some tags to our cloud builds (#94)
This allows us to find builds a little easier in the GCB history dashboard. It has a lot of useful filters but they're pretty much all based on the CI trigger. Since we don't trigger from a commit, we can't really use any of them. But we can use tags!