feat(gcb): use GCB to build the published debs
When we publish Halyard, we have to rebuild the debs again. I hadn't converted that, so it was still happening on the Jenkins server instead of in cloud build. And now it's failing because we didn't apply #98 to the gradle runner code... which was fortunate because it made me realize I need to convert this code over to GCB.