fix(kubernetes): deploy test Spinnaker using a V2 account (#125)
* feat(kubernetes): run distributed install on kubernetes v2 (#32)

* feat(kubernetes): run distributed install on kubernetes v2

* fix(kubernetes): use label selectors for pod name lookup

* fix(kubernetes): Fix race condition in V2 deployment (#33)

When deploying using the V2 provider, control is ceded back to the
CI tests once the deployments have been applied, whether or not they
are stable yet.  This is causing errors on setting up port forwarding,
as the pods are sometimes not running when we try to forward to them.
To address this, let's block on deployment availability before trying
to port forward.

* fix(kubernetes): Actually fix race condition in V2 deployment (#35)

The last commit added some great logic to make sure that V2
deployments are ready before we try to port-forward to them,
except that it added the logic to the V1 deployer instead of
the V2 deployer. Move this logic to the V2 deployer, and add
a couple of log lines so it's easier to see that it's working.

Co-authored-by: Eric Zimanyi <ezimanyi@google.com>