fix(kubernetes): Delete broken V1 test (#139)
This test does not work with Kubernetes 1.16, which is what our
test cluster is not upgraded to. Given that the V1 provider is
deprecated, it's not worth the significant effort to figure out
how this test is broken.  (It's almost certainly something in the
test, though even if the V1 provider were actually broken deploying
to 1.16 we would not fix it given our deprecation policy.)

Obviously if anyone wants to debug this and can fix it, we can
restore it, but for now I'm deleting it to unblock further
patch releases on the 1.19 and 1.20 branches.

(cherry picked from commit 47c80467b3585d396e635a6ee320191a6d962072)