fix(audit): Fix a few issues with Admin_AuditBoms (#152)
* fix(audit): Remove unused keep_latest_version flag

We are not using this flag; removing it allows us to remove some
unecessary code, some of which is throwing a lot of exceptions that
clutter the logs.

* fix(audit): Don't require a jar for deck

At some point we stopped publishing jars for deck. (I'm not sure
what was even in them when we did...)  But we should not report
a failure if a released deck version doesn't have a jar.

* fix(audit): Don't remove versions whose prefix matches

Now that we publish variants of containters with -java8 or
-slim or -slim-java8 at the end, we should not require the
tag to exactly match a version we released, but should just
start with such a version.

* fix(audit): Reduce concurrency of gsutil

It seems like gsutil itself has locking and limits to only one
operation running at a time, so trying to run 64 of these at once
is not any faster.  We'll still use 2 threads in case we speed things
up a tiny bit by letting the next job run while we're parsing the
yaml from the prior one.

* fix(audit): Update concurrency to 4

Actually after running this a few times, up to parallel 4 does
speed things up, so let's use that instead of 2.