fix(test): fix kube_v2_cache_test
The behavior previously enabled by the `liveManifestCalls` flag is now [enabled by default](https://github.com/spinnaker/spinnaker/issues/5906). We have removed the Force Cache Refresh task from all Kubernetes pipelines stages (these were a no-op with the flag enabled). However, this resulted in the end-to-end tests that expected the cache to be in a consistent state after Kubernetes operations completed to begin failing intermittently.

This commit adds a retry interval of 8 seconds with a max retries of 15, for a total of 2 minutes, to all test cases that expect the cache to reflect the state of the operation being tested. This includes any operation that hits Clouddriver's applications, clusters, server group, and load balancers endpoints. It excludes those, such as `delete_kind`, that instead rely on the citest.kube_testing package. The retry interval and max retries numbers were chosen for consistency with the existing test that relies on a [cache update](https://github.com/spinnaker/buildtool/blob/master/testing/citest/tests/kube_v2_cache_test.py#L412).