fix(containers): fix checking for existing containers (#162)
Listing images with a filter doesn't seem to work. It seems to only find the image we're looking for if it was uploaded recently. I suspect it just does a client-side filter of the first page of results or something. Since we know exactly what tag we want anyway, let's just look for that specific one.

Right now on old release branches, we end up rebuilding containers that already exist because we don't find them. This should take care of that.

(cherry picked from commit ee071a832ccc829433200c7a0fbb37dfa5253b83)