chore(ci): Publish Changelog to spinnaker.io without gists
Each `changelog.md` file only includes changes for a single release
version (patch).

Previously, `buildtool` would concatenate the gist URL for each previous patch
version for the same major.minor. e.g: 1.2.3 would include gists for
1.2.3, 1.2.2, 1.2.1, 1.2.0 releases.

This can be nice to view on the page but we can potentially handle
this with Hugo instead if it's necessary.