feat(ci): Tidy Publish Spinnaker - remove imports, dead code
Previously we did a two stage process of tagging branches, then pushing.
Maybe this was related to the BOM and changelog generation process but
these days with GHA we don't have a concern here I think.

For example:
1. we tag Deck release-XYZ branch and push tag
2. failure occurs
3. buildtool is re-run, it will see Deck is already tagged at HEAD of
   release-XYZ branch and skip tagging, continuing on to next
   repository

We also previously checked for tags in a repo. This isn't necessary
either as we don't have any automatic recourse available.
If the tag already exists in the repo then a git push will fail and this
will bubble up and cause buildtool to exit with a non-zero status.
Someone can manually investigate why.