feat(ci): Publish Spinnaker command (#196)
* feat(ci): Publish Spinnaker - tag & create branches

- tag branches in our bumpdep order, master or release-* depending on
  new release version
- create new release-* branches if required

* feat(ci): Tidy Publish Spinnaker - remove imports, dead code

Previously we did a two stage process of tagging branches, then pushing.
Maybe this was related to the BOM and changelog generation process but
these days with GHA we don't have a concern here I think.

For example:
1. we tag Deck release-XYZ branch and push tag
2. failure occurs
3. buildtool is re-run, it will see Deck is already tagged at HEAD of
   release-XYZ branch and skip tagging, continuing on to next
   repository

We also previously checked for tags in a repo. This isn't necessary
either as we don't have any automatic recourse available.
If the tag already exists in the repo then a git push will fail and this
will bubble up and cause buildtool to exit with a non-zero status.
Someone can manually investigate why.

* feat(ci): Publish Spinnaker command - build bom & changelog

* feat(ci): Fetch BOMs from GCS bucket instead of Halyard

We have been replacing Halyard specific functionality with native
buildtool python or shell calls. This is another one.

There was already a dependency on `gsutil` for GCS bucket interactions
within buildtool.

* feat(ci): Publish Spinnaker - add publish changelog

* feat(ci): Add Publish BOM command

* feat(ci): Publish Spinnaker command - update versions.yml

* chore(ci): Removed unused get_next_patch_parameters command

Whilst it would be nice for release orchestration to happen
automatically, for now we are deciding what version to do up front and
supplying this to buildtool. Ideally at some point via a git tag & push.

* chore(ci): Update file comments

* feat(ci): Publish Spinnaker command - default dry_run

Validate in test.

* feat(ci): Publish Spinnaker command - README

* chore(ci): Disable tagging & branch creation

Tagging some repo's like kork result in other repo's receiving commits
and new tags. This all takes time.

Until this is solved for we can't tag repo's.

We can technically still branch as buildtool will walk back to the last
tag on master. For now disable so no surprises.

We also need container images to be built and pushed to GAR before we
can add additional spinnaker release tags. This requires bumpdep/etc to
propagate.

* chore(ci): README next release, clarify arg param