feat(ci): Change Release to manual trigger only
`publish_spinnaker` command requires a minimum halyard version. It could
be optional but we still need a way to bump it from time to time.
This makes using a pure git tag `v1.30.0` flow awkward.

Go all in on using GitHub Actions "workflow dispatch" manual action
triggering where we can support as many parameters as we need and
default them for safety.