Provide creds (#202)
* feat(gha): add new --github_oauth_token option

and supply a token, so we have permission to push a branch/PR to spinnaker.io containing
the changelog.

Note that GHA obscures the token in the log/job output so it's not visible.

* feat(gha): force push the changelog branch to spinnaker.io

in case the branch already exists (e.g. from a previous failed release attempt)

* feat(gha): use gcloud creds so gsutil cp has a chance to publish the bom and versions files

apologies for the duplicate code