feat(regctl): Support tagging containers with regctl
`regctl` supports tagging containers directly in a container registry without
doing a pull/tag/push flow. This is great because it saves a lot of time
with network transfers.

When cutting a Spinnaker release we need to tag slim (Alpine) and
Ubuntu containers at a specific <tag>:
- without `-unvalidated`
- with our `spinnaker-<release_version>`.

Default to `--dry_run true` mode to prevent tagging containers in the
registry by mistake.